+++ env/dev/clj/user.clj
@@ -3,32 +3,38 @@
 
   Usage:
   o (reset) - [re]run the app.
-  o In case REPL fails after syntax error, call (refresh) and try again.
   "
   (:require [pickings.core :as core]
-            [com.stuartsierra.component :as component]
-            [clojure.tools.namespace.repl :refer [refresh]]
+            [integrant.core :as ig]
+            [clojure.tools.namespace.repl :refer [set-refresh-dirs refresh refresh-all]]
             [clojure.repl :refer :all]
             [clojure.pprint :refer :all]))
 
-;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;; CLJ
+(println "Hi dev!")
+
 (def system (atom nil))
 
-(println "Hi!")
+(defn -start
+  []
+  ; pass existing system because key provider can be already initialized and cannot be stopped via REPL (it's a bug of keymaster lib)
+  (swap! system (comp ig/init core/new-config)))
 
-(defn- init []
-  (swap! system (constantly (core/new-system))))
+(defn -stop
+  []
+  (when @system (ig/halt! @system)))
 
-(defn- start []
-  (swap! system component/start))
+(defn reset
+  []
+  (-stop)
 
-(defn- stop []
-  (swap! system (fn [s] (when s (component/stop s)))))
+  ; this will enforce reloading of the methods and detection of missing methods
+  (remove-all-methods ig/init-key)
+  (remove-all-methods ig/halt-key!)
+  ; TODO: add other methods and/or see https://github.com/weavejester/integrant/issues/6 and https://github.com/weavejester/integrant/issues/8
 
-(defn- go []
-  (init)
-  (start))
+  ; do not refresh this ns because it stores the previous system value needed on restart
+  (set-refresh-dirs "src")
 
-(defn reset []
-  (stop)
-  (refresh :after 'user/go))
+  ; refresh-all is used instead of refresh because we've just removed all the methods and want to find all the definitions
+  (refresh-all :after 'user/-start)
+  nil)